jekyll: Move theme to _config-dev
If you have a `theme` in `_config.yml` you'll get nagged by GitHub
on every pages build that it's not supported.

However, when running locally I really don't want to use the remote
theme, I have no need for it and it's annoying when offline. Buuuuuut
regardless of `theme` the `remote-theme` gem just won't take a hint. The
idea was to set `remote_theme` to `null` to unset it but that doesn't
work since you can't unset shit, only override it. So instead it's now
set to empty string, which triggers a big red warning when building locally
b/c that's not a valid repo/ref but at least works.